picturas-change-brightness-tool-ms-main/pyproject.toml
[tool.poetry]
name = "picturas-change-brightness-tool-ms"
version = "0.1.0"
description = "PictuRAS microservice (tool) that applies a watermark to an input image, generating a new watermarked output image."
authors = ["Afonso Costa <d13690@di.uminho.pt>", "Paulo Sousa <d7157@di.uminho.pt>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pika = "^1.3.2"
pydantic = "^2.10.3"
pillow = "^11.0.0"


[tool.poetry.group.dev.dependencies]
yapf = "^0.43.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
